AudioService: added logic to push currently playing Rooster to next alarm if dismissed midway. TODO: allow skipping?

+++ app/src/main/java/com/roostermornings/android/background/AudioService.java
@@ -194,6 +194,7 @@ public class AudioService extends Service {
     }
 
     public void endService(ServiceConnection conn) {
+        if(mediaPlayerRooster != null && mediaPlayerRooster.isPlaying()) return;
         try {
             file.delete();
         } catch(NullPointerException e){
